Sesión03/lab03-fastapi-grpc-todo/services/todo-api/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
pydantic